refactor: views becomes a package, one module per screen
views.py went from a hundred lines to seven hundred in a day, and every
feature since has had to land in it — a file two people cannot both work
in. It splits the way n23's did: a package, one module per screen, named
for the pipeline module each faces, so views.hire is the web face of
core.hire and where a view lives is a thing to guess rather than
remember. permissions.py holds the two owner guards both halves share.

A pure move: every function body is carried across byte-for-byte and the
package re-exports what urls.py reads, so nothing outside changed. The
suite is 1397 passing either side of it.
